+++ src/com/genir/aitweaks/core/features/shipai/autofire/SyncState.kt
@@ -1,0 +1,4 @@
+package com.genir.aitweaks.core.features.shipai.autofire
+
+/** SyncState is used to synchronise weapons that are to fire in staggered mode. */
+data class SyncState(var weapons: Int, var lastAttack: Float)